requirements.txt
#
annotated-doc==0.0.3; python_version >= '3.8'
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.11.0; python_version >= '3.9'
certifi==2025.10.5; python_version >= '3.7'
chardet==5.2.0; python_version >= '3.7'
charset-normalizer==3.4.4; python_version >= '3.7'
click==8.3.0; python_version >= '3.10'
coloredlogs==15.0.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
contourpy==1.3.3; python_version >= '3.11'
cycler==0.12.1; python_version >= '3.8'
fastapi==0.120.1; python_version >= '3.8'
httpx==0.28.1; python_version >= '3.8'
flatbuffers==25.9.23
fonttools==4.60.1; python_version >= '3.9'
h11==0.16.0; python_version >= '3.8'
humanfriendly==10.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
idna==3.11; python_version >= '3.8'
joblib==1.5.2; python_version >= '3.9'
jsonpickle==4.1.1; python_version >= '3.8'
kiwisolver==1.4.9; python_version >= '3.10'
matplotlib==3.10.7; python_version >= '3.10'
midi2audio==0.1.1
more-itertools==10.8.0; python_version >= '3.9'
mpmath==1.3.0
music21==9.9.1; python_version >= '3.10'
numpy==2.2.6; python_version >= '3.10'
oemer==0.1.5
onnxruntime==1.23.2; python_version >= '3.10'
opencv-python==4.12.0.88; python_version >= '3.6'
packaging==25.0; python_version >= '3.8'
pillow==12.0.0; python_version >= '3.10'
protobuf==6.33.0; python_version >= '3.9'
pydantic==2.12.3; python_version >= '3.9'
pydantic-core==2.41.4; python_version >= '3.9'
pyparsing==3.2.5; python_version >= '3.9'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
python-multipart==0.0.18
pytest==8.3.3
requests==2.32.5; python_version >= '3.9'
scikit-learn==1.7.2; python_version >= '3.10'
scipy==1.16.2; python_version >= '3.11'
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
sniffio==1.3.1; python_version >= '3.7'
starlette==0.49.0; python_version >= '3.9'
sympy==1.13.1; python_version >= '3.9'
torch==2.6.0; python_version >= '3.9'
torchvision==0.21.0; python_version >= '3.9'
threadpoolctl==3.6.0; python_version >= '3.9'
typing-extensions==4.15.0; python_version >= '3.9'
typing-inspection==0.4.2; python_version >= '3.9'
urllib3==2.5.0; python_version >= '3.9'
uvicorn==0.38.0; python_version >= '3.9'
webcolors==24.11.1; python_version >= '3.9'
